Add search and install commands
