+++ src/dotnet-bootstrapper/BootstrapperCommandParser.cs
@@ -8,7 +8,8 @@ using System.CommandLine.Invocation;
 using System.CommandLine.Parsing;
 using System.Reflection;
 using Microsoft.DotNet.Tools.Uninstall.Shared.Configs;
-using Microsoft.DotNet.Tools.Bootstrapper.Commands.List;
+using Microsoft.DotNet.Tools.Bootstrapper.Commands.Search;
+using Microsoft.DotNet.Tools.Bootstrapper.Commands.Install;
 
 namespace Microsoft.DotNet.Tools.Bootstrapper
 {
@@ -22,7 +23,8 @@ namespace Microsoft.DotNet.Tools.Bootstrapper
         static BootstrapperCommandParser()
         {
             BootstrapperRootCommand.AddCommand(CommandLineConfigs.VersionSubcommand);
-            BootstrapperRootCommand.AddCommand(ListCommandParser.GetCommand());
+            BootstrapperRootCommand.AddCommand(SearchCommandParser.GetCommand());
+            BootstrapperRootCommand.AddCommand(InstallCommandParser.GetCommand());
             BootstrapperRootCommand.AddCommand(CommandLineConfigs.RemoveCommand);
             BootstrapperRootCommand.AddCommand(HelpCommand);
             HelpCommand.Handler = CommandHandler.Create(() =>
